requirements.txt
tensorflow-gpu==1.12.0
gensim
sklearn
numpy